+++ myqueue/states.py
@@ -34,7 +34,7 @@ class State(Enum):
     >>> State.str2states('dA')
     """
 
-    UNDEFINED = 'U'
+    undefined = 'u'
     queued = 'q'
     hold = 'h'
     running = 'r'
@@ -60,7 +60,7 @@ class State(Enum):
         return self.name
 
     def is_bad(self) -> bool:
-        """Return true for UNDEFINED, FAILED, TIMEOUT, MEMORY and CANCELED.
+        """Return true for FAILED, TIMEOUT, MEMORY and CANCELED.
 
         >>> State.running.is_bad()
         False