+++ myqueue/test/test_workflow.py
@@ -1,11 +1,13 @@
 import os
 from pathlib import Path
+
 from myqueue.task import task
 from myqueue.test.flow1 import workflow
 from myqueue.test.hello import workflow as hello
 
 
 def create_tasks():
+    """Old style."""
     return [task('somepackage.somemodule')]
 
 
@@ -45,3 +47,16 @@ def test_direct_hello(tmp_path):
     os.chdir(tmp_path)
     Path('hello.sh').write_text('echo $@')
     hello()
+
+
+def test_flow2(mq):
+    script = Path(__file__).with_name('flow2.py')
+    mq(f'workflow {script}')
+    mq.wait()
+    assert mq.states() == 'MCCC'
+    mq('rm -sC .')
+    mq(f'workflow {script}')
+    assert mq.states() == 'M'
+    mq(f'workflow {script} --force')
+    mq.wait()
+    assert mq.states() == 'MCCC'